Change user model for floats and fix address bar coords in create listings

+++ server/user/userModel.js
@@ -10,7 +10,7 @@ var User = db.define('user', {
   rating: Sequelize.INTEGER,
   picReference: Sequelize.STRING,
   zipcode: Sequelize.INTEGER,
-  avgRating: { type: Sequelize.INTEGER, allowNull: false, defaultValue: 0 },  //
+  avgRating: { type: Sequelize.FLOAT, allowNull: false, defaultValue: 0 },  //
   ratingCount: { type: Sequelize.INTEGER, allowNull: false, defaultValue: 0 },  //
 });
